requirements.txt
asgiref==3.7.2
mysqlclient==2.2.0
PyMySQL==1.1.0
setuptools==65.5.1
sqlparse==0.4.4
tzdata==2023.3
wheel==0.38.4
django-jazzmin>=2.4.6,<3.0
django-import-export>=2.7.0,<3.0

# Add other dependencies as needed